+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 0.11.0 (TBD)
 
 - [BREAKING] Incremented MSRV to 1.88.
+- [BREAKING] Added payback `NoteType` parameter to standard SWAP note (#1539).
+
 
 ## 0.10.0 (2025-07-08)